+++ lib/rubocop/cop/rspec/scattered_setup.rb
@@ -7,6 +7,14 @@ module RuboCop
       #
       # Unify `before`, `after`, and `around` hooks when possible.
       #
+      # Autocorrection is not supported when `around` is used because it can
+      # be difficult to unify them.
+      #
+      # @safety
+      #   This cop's autocorrection is unsafe because it could change the order
+      #   of processing if there is another hook between them.
+      #   (e.g. `let!` between `before`s)
+      #
       # @example
       #   # bad
       #   describe Foo do
@@ -23,45 +31,71 @@ module RuboCop
       #   end
       #
       class ScatteredSetup < Base
+        extend AutoCorrector
+
+        include RangeHelp
+
         MSG = 'Do not define multiple `%<hook_name>s` hooks in the same ' \
-              'example group (also defined on %<lines>s).'
+              'example group (also defined on line %<line>s).'
 
         def on_block(node) # rubocop:disable InternalAffairs/NumblockHandler
-          return unless example_group?(node)
+          return unless hook?(node)
 
-          repeated_hooks(node).each do |occurrences|
-            lines = occurrences.map(&:first_line)
+          same_hook = find_same_hook(node)
+          return unless same_hook
 
-            occurrences.each do |occurrence|
-              lines_except_current = lines - [occurrence.first_line]
-              message = format(MSG, hook_name: occurrences.first.method_name,
-                                    lines: lines_msg(lines_except_current))
-              add_offense(occurrence, message: message)
-            end
+          add_offense(node, message: format_message(same_hook)) do |corrector|
+            next if node.method?(:around)
+
+            merge_hooks(corrector, same_hook, node)
           end
         end
 
         private
 
-        def repeated_hooks(node)
-          hooks = RuboCop::RSpec::ExampleGroup.new(node)
-            .hooks
-            .select(&:knowable_scope?)
-            .group_by { |hook| [hook.name, hook.scope, hook.metadata] }
-            .values
-            .reject(&:one?)
+        def find_same_hook(node)
+          return unless hook_knowable_scope(node)
 
-          hooks.map do |hook|
-            hook.map(&:to_node)
+          node.left_siblings.find do |sibling|
+            hook?(sibling) &&
+              sibling.method?(node.method_name) &&
+              hook_knowable_scope(node) &&
+              hook_scope(sibling) == hook_scope(node) &&
+              hook_metadata(sibling) == hook_metadata(node)
           end
         end
 
-        def lines_msg(numbers)
-          if numbers.size == 1
-            "line #{numbers.first}"
-          else
-            "lines #{numbers.join(', ')}"
-          end
+        def format_message(node)
+          format(
+            MSG,
+            hook_name: node.method_name,
+            line: node.location.line
+          )
+        end
+
+        def hook_knowable_scope(node)
+          RuboCop::RSpec::Hook.new(node).knowable_scope?
+        end
+
+        def hook_metadata(node)
+          RuboCop::RSpec::Hook.new(node).metadata
+        end
+
+        def hook_scope(node)
+          RuboCop::RSpec::Hook.new(node).scope
+        end
+
+        def merge_hooks(corrector, node1, node2)
+          corrector.insert_after(
+            node1.body || node1.location.begin,
+            "\n\n#{node2.body.source}"
+          )
+          corrector.remove(
+            range_by_whole_lines(
+              node2.location.expression,
+              include_final_newline: true
+            )
+          )
         end
       end
     end