+++ spec/rubocop/cop/rspec/scattered_setup_spec.rb
@@ -2,42 +2,65 @@
 
 RSpec.describe RuboCop::Cop::RSpec::ScatteredSetup do
   it 'flags multiple hooks in the same example group' do
-    expect_offense(<<-RUBY)
+    expect_offense(<<~RUBY)
       describe Foo do
         before { bar }
-        ^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on line 3).
         before { baz }
         ^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on line 2).
+      end
+    RUBY
+
+    expect_correction(<<~RUBY)
+      describe Foo do
+        before { bar
+
+      baz }
       end
     RUBY
   end
 
   it 'flags multiple hooks of the same scope with different symbols' do
-    expect_offense(<<-RUBY)
+    expect_offense(<<~RUBY)
       describe Foo do
         after { bar }
-        ^^^^^^^^^^^^^ Do not define multiple `after` hooks in the same example group (also defined on lines 3, 4).
         after(:each) { baz }
-        ^^^^^^^^^^^^^^^^^^^^ Do not define multiple `after` hooks in the same example group (also defined on lines 2, 4).
+        ^^^^^^^^^^^^^^^^^^^^ Do not define multiple `after` hooks in the same example group (also defined on line 2).
         after(:example) { baz }
-        ^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `after` hooks in the same example group (also defined on lines 2, 3).
+        ^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `after` hooks in the same example group (also defined on line 2).
+      end
+    RUBY
+
+    expect_correction(<<~RUBY)
+      describe Foo do
+        after { bar
+
+      baz
+
+      baz }
       end
     RUBY
   end
 
   it 'flags multiple before(:all) hooks in the same example group' do
-    expect_offense(<<-RUBY)
+    expect_offense(<<~RUBY)
       describe Foo do
         before(:all) { bar }
-        ^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on line 3).
         before(:all) { baz }
         ^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on line 2).
+      end
+    RUBY
+
+    expect_correction(<<~RUBY)
+      describe Foo do
+        before(:all) { bar
+
+      baz }
       end
     RUBY
   end
 
   it 'ignores different hooks' do
-    expect_no_offenses(<<-RUBY)
+    expect_no_offenses(<<~RUBY)
       describe Foo do
         before { bar }
         after { baz }
@@ -47,7 +70,7 @@ RSpec.describe RuboCop::Cop::RSpec::ScatteredSetup do
   end
 
   it 'ignores different hook types' do
-    expect_no_offenses(<<-RUBY)
+    expect_no_offenses(<<~RUBY)
       describe Foo do
         before { bar }
         before(:all) { baz }
@@ -57,7 +80,7 @@ RSpec.describe RuboCop::Cop::RSpec::ScatteredSetup do
   end
 
   it 'ignores hooks in different example groups' do
-    expect_no_offenses(<<-RUBY)
+    expect_no_offenses(<<~RUBY)
       describe Foo do
         before { bar }
 
@@ -69,7 +92,7 @@ RSpec.describe RuboCop::Cop::RSpec::ScatteredSetup do
   end
 
   it 'ignores hooks in different shared contexts' do
-    expect_no_offenses(<<-RUBY)
+    expect_no_offenses(<<~RUBY)
       describe Foo do
         shared_context 'one' do
           before { bar }
@@ -83,7 +106,7 @@ RSpec.describe RuboCop::Cop::RSpec::ScatteredSetup do
   end
 
   it 'ignores similar method names inside of examples' do
-    expect_no_offenses(<<-RUBY)
+    expect_no_offenses(<<~RUBY)
       describe Foo do
         before { bar }
 
@@ -95,7 +118,7 @@ RSpec.describe RuboCop::Cop::RSpec::ScatteredSetup do
   end
 
   it 'ignores hooks with different metadata' do
-    expect_no_offenses(<<-RUBY)
+    expect_no_offenses(<<~RUBY)
       describe Foo do
         before(:example) { foo }
         before(:example, :special_case) { bar }
@@ -104,18 +127,75 @@ RSpec.describe RuboCop::Cop::RSpec::ScatteredSetup do
   end
 
   it 'flags hooks with similar metadata' do
-    expect_offense(<<-RUBY)
+    expect_offense(<<~RUBY)
       describe Foo do
         before(:each, :special_case) { foo }
-        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on lines 3, 4, 5).
         before(:example, :special_case) { bar }
-        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on lines 2, 4, 5).
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on line 2).
         before(:example, special_case: true) { bar }
-        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on lines 2, 3, 5).
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on line 2).
         before(special_case: true) { bar }
-        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on lines 2, 3, 4).
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on line 2).
         before(:example, special_case: false) { bar }
       end
     RUBY
+
+    expect_correction(<<~RUBY)
+      describe Foo do
+        before(:each, :special_case) { foo
+
+      bar
+
+      bar
+
+      bar }
+        before(:example, special_case: false) { bar }
+      end
+    RUBY
+  end
+
+  context 'with multiple `around`' do
+    it 'registers offense without autocorrection' do
+      expect_offense(<<~RUBY)
+        describe Foo do
+          around do |example|
+            a
+            example.run
+            b
+          end
+
+          around do |example|
+          ^^^^^^^^^^^^^^^^^^^ Do not define multiple `around` hooks in the same example group (also defined on line 2).
+            c
+            example.run
+            d
+          end
+        end
+      RUBY
+
+      expect_no_corrections
+    end
+  end
+
+  context 'with `let!` between multiple `before`' do
+    it 'registers offense' do
+      expect_offense(<<~RUBY)
+        describe Foo do
+          before { a }
+          let!(:b) { 1 }
+          before { c }
+          ^^^^^^^^^^^^ Do not define multiple `before` hooks in the same example group (also defined on line 2).
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        describe Foo do
+          before { a
+
+        c }
+          let!(:b) { 1 }
+        end
+      RUBY
+    end
   end
 end